Mostramos una lista con datos extraídos de Firebase

+++ app/src/main/java/helicida/firebasetest/User.java
@@ -6,17 +6,20 @@ import com.firebase.client.Firebase;
  * Created by 46465442z on 27/01/16.
  */
 public class User {
+
     private int birthYear;
     private String fullName;
 
-    public User() {}
+    public User() {
+
+    }
 
     public User(String fullName, int birthYear) {
         this.fullName = fullName;
         this.birthYear = birthYear;
     }
 
-    public long getBirthYear() {
+    public int getBirthYear() {
         return birthYear;
     }
